Introduce a package-private BaseGraph interface and AbstractBaseGraph class. This allows us to remove a large amount of code duplication. It also gives us a small reduction in constant overhead for Graphs.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=144926159
