+++ guava/src/com/google/common/graph/ImmutableValueGraph.java
@@ -66,7 +66,7 @@ public final class ImmutableValueGraph<N, V> extends ConfigurableValueGraph<N, V
 
   @Override
   public ImmutableGraph<N> asGraph() {
-    return new ImmutableGraph<N>(super.asGraph()); // safe because the view is effectively immutable
+    return new ImmutableGraph<N>(this); // safe because the view is effectively immutable
   }
 
   private static <N, V> ImmutableMap<N, GraphConnections<N, V>> getNodeConnections(
